[routes] Load routes from included router
